feat: implement task creation and improve navigation between trip tabs #22 sopra-fs26-group-29-server#13
